requirements.txt
streamlit==0.63.0
pytube3==9.6.4
Pillow==7.1.2
hurry.filesize==0.9
requests==2.23.0